% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp.R
\name{calc_pve}
\alias{calc_pve}
\title{Calculate PVE from sd of u and p/n}
\usage{
calc_pve(sigu, p_n)
}
\arguments{
\item{sigu}{sd of u}

\item{p_n}{p/n}
}
\value{

}
\description{
Calculate PVE from sd of u and p/n
}
