% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp.R
\name{calc_varu}
\alias{calc_varu}
\title{calculate variance of u from pve and p_n}
\usage{
calc_varu(pve, p_n)
}
\arguments{
\item{pve}{pve}

\item{p_n}{p/n}
}
\description{
calculate variance of u from pve and p_n
}
