% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp_utils.R
\name{convert_quh}
\alias{convert_quh}
\title{Convert summary statistics to PC-space}
\usage{
convert_quh(uhat, Q)
}
\arguments{
\item{uhat}{a vector of length `p` containing marginal association Z scores.  Alternatively, a  matrix with `p` rows, each column
representing marginal association with a different trait}

\item{Q}{A `q` by `p` matrix consisting of `q` eigenvectors of the LD matrix computed at p loci.}
}
\description{
Convert summary statistics to PC-space
}
