% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{est_sim}
\alias{est_sim}
\title{Estimate parameters from simulated data, saved in the specified format}
\usage{
est_sim(resl, Ql = NULL, D = NULL, doConfound = T, log_params = F,
  useGradient = T, bias_bounds = c(0, 0.3), pve_bounds = c(1e-04, 1))
}
\arguments{
\item{resl}{list with the following elements 
`quh_mat`, a `p`x`g` matrix (where `p` is number of SNPs and `g` is number of traits) obtained by an
operation equivalent to to `crossprod(Q,uh)`
`tparam_df` a dataframe (of the type generated by the function `gen_tparamdf_norm` with true parameter values (`NA`s can also be used in the case of real data)}

\item{Ql}{A list of matrices, each element representing the eigenvectors of a square block in the block diagonal approximation of the LD matrix (blocks need not be of equal size)}

\item{D}{A vector of the eigenvectors of the LD matrix (must be length `p`)}

\item{doConfound}{a logical indicating whether to use the two parameter model (`doConfound=T`) or the one parameter model without confounding}

\item{log_params}{a logical indicating whether to optimize parameters in log space or not (not recommended)}

\item{a_bounds}{a length two vector indicating the bounds of the parameter `a` to use when optimizing, `a` is a measure of confounding}

\item{sigu_bounds}{a length two vector indicating bounds in the parameter `sigu` to use when optimizing `sigu` is directly related to PVE}
}
\description{
Estimate parameters from simulated data, saved in the specified format
}
