% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp_utils.R, R/RcppExports.R
\name{estimate_pve}
\alias{estimate_pve}
\title{Function for obtaining pve estimate from the RSSp optimizer}
\usage{
estimate_pve(cvec, D, quh, sample_size)

estimate_pve(cvec, D, quh, sample_size)
}
\arguments{
\item{cvec}{an estimate of the variance}

\item{quh}{uhat a vector of length `p` containing marginal association Z scores transformed into PC space.}

\item{dvec}{uhat a vector of length `p` containing eigenvalues}

\item{N}{the sample size}
}
\description{
Function for obtaining pve estimate from the RSSp optimizer

compute pve with confounding
}
