% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evd_dnorm_grad_stan}
\alias{evd_dnorm_grad_stan}
\title{evd_dnorm_grad_stan}
\usage{
evd_dnorm_grad_stan(par, dvec, quh)
}
\arguments{
\item{quh}{The precomputed matrix vector product Q%*%u_hat (passed as a vector)}

\item{dvec.}{The eigenvalues of the LD matrix}
}
\description{
This is an attempt to use stan's AD features to calculate a gradient 
for the RSSp likelihood
}
\details{
//' @param par a length 2 numeric vector.  par[1] corresponds to the sigma_u parameter, and par[2] corresponds to the confounding parameter
}
