% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evd_post_var}
\alias{evd_post_var}
\title{Compute the posterior variance of `crossprod(Q,u)` given eigenvalues, `crossprod(Q,u_hat)`,sigma_u^2 and confounding levels.}
\usage{
evd_post_var(par, dvec)
}
\arguments{
\item{par}{a length 2 numeric vector.  `par[1]` corresponds to the sigma_u^2 parameter, and `par[2]` corresponds to the confounding parameter. 
If the model does not include confounding, `par` can be a length one vector.}

\item{dvec.}{The eigenvalues of the LD matrix, passed as a length `p` vector}

\item{quh}{The precomputed matrix vector product `crossprod(Q,u_hat)` (passed as a  length `p` vector)}
}
\description{
Compute the posterior variance of `crossprod(Q,u)` given eigenvalues, `crossprod(Q,u_hat)`,sigma_u^2 and confounding levels.
}
