% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gen_ti}
\alias{gen_ti}
\title{calculate the scale of residuals, given X*Beta and the target PVE}
\usage{
gen_ti(vy, PVE)
}
\arguments{
\item{vy}{vector with the variance of y (length is equal to the number of traits)}

\item{PVE}{vector specifying the target PVE (length should be the same as the number of traits)}
}
\description{
calculate the scale of residuals, given X*Beta and the target PVE
}
