% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{gen_tparamdf_norm}
\alias{gen_tparamdf_norm}
\title{Generate parameters of simulation}
\usage{
gen_tparamdf_norm(pve, bias = 0, nreps, n, p, fgeneid = NULL)
}
\arguments{
\item{pve}{true PVE}

\item{bias}{true value for bias/confounding}

\item{nreps}{number of repetitions of each combo of parameters}

\item{n}{sample size}

\item{p}{number of SNPs}

\item{fgeneid}{name for each trait (defaults to 1:ntraits)}
}
\description{
Generate parameters of simulation
}
