% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp_utils.R
\name{gensplit_uh_uf}
\alias{gensplit_uh_uf}
\title{Function factory for obtaining PC-transformed summary statistics}
\usage{
gensplit_uh_uf(uh, region_id)
}
\arguments{
\item{uh}{uhat a vector of length `p` containing marginal association Z scores.  Alternatively, a  matrix with `p` rows, each column
representing marginal association with a different trait}

\item{chunklist}{a vector of length p indicating which chunk each locus is assigned to}
}
\description{
Function factory for obtaining PC-transformed summary statistics
}
