% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp_utils.R
\name{map_convert_quh}
\alias{map_convert_quh}
\title{Convert chunks of summary statistics to PC-space}
\usage{
map_convert_quh(chunklist, uhat_uf, q_uf, map_fun = purrr::map, ...)
}
\arguments{
\item{chunklist}{a vector or list specifying LD regions. each elemenent of chunklist will be passed to `uhat_uf` and `q_uf`}

\item{uhat_uf}{a function that takes a single element of `chunklist` and returns either a length p vector, or a matrix with p rows}

\item{Q_uf}{a function that takes a single element of `chunklist` and returns a `q` by `p` matrix}
}
\description{
Convert chunks of summary statistics to PC-space
}
