% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp.R
\name{modify_R}
\alias{modify_R}
\title{Modify LD matrix}
\usage{
modify_R(R, z, ref_sample_size)
}
\arguments{
\item{R}{LD matrix}

\item{z}{vector of z scores}

\item{ref_sample_size}{reference panel size}
}
\value{
matrix of same dimension as D
}
\description{
Modify LD matrix
}
