% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panel_eigenvalues}
\alias{panel_eigenvalues}
\title{Sample LD matrix eigenvalues}
\format{
a vector with 312 elements with sorted eigenvalues
}
\usage{
panel_eigenvalues
}
\description{
A vector of eigenvalues from a region of 1000 genomes EUR individuals
}
\keyword{datasets}
