% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{parse_ldsc_h2log}
\alias{parse_ldsc_h2log}
\title{read log from LD score regression (as implemented in `ldsc`), and parse it so results can be compared to RSSp}
\usage{
parse_ldsc_h2log(h2lf)
}
\arguments{
\item{h2lf}{path (absolute or relative) to `ldsc` (character)}
}
\description{
read log from LD score regression (as implemented in `ldsc`), and parse it so results can be compared to RSSp
}
