% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp_utils.R
\name{post_var_D}
\alias{post_var_D}
\title{Estimate the variance of each eigensnp}
\usage{
post_var_D(cvec, D, ind = seq_along(cvec))
}
\arguments{
\item{cvec}{inferred parameters for the model}

\item{D}{vector of eigenvalues of the LD matrix}

\item{ind}{indices}
}
\value{
Nicholas Knoblauch \email{nwknoblauch@gmail.com}
}
\description{
Estimate the variance of each eigensnp
}
