% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp.R
\name{prep_RSSp}
\alias{prep_RSSp}
\title{Prepare data for RSSp by performing eigenvalue decomposition on a list of LD matrices,
and transforming GWAS summary statistics based on the decomposition}
\usage{
prep_RSSp(Rl, U, N)
}
\arguments{
\item{Rl}{A list of symmetric positive-semidefinite matrices, with one matrix for each LD region}

\item{U}{A vector or matrix of GWAS standardized effect sizes. If `U` is a matrix, there should be one column for each trait.
It is generally a good idea to keep track of the traits by naming the columns}
}
\value{
Returns a dataframe with columns `quh`, which is the transformed effect size,`D` which are the eigenvalues and `fgeneid`
which is the trait name (taken from the column names of `U`)
}
\description{
Prepare data for RSSp by performing eigenvalue decomposition on a list of LD matrices,
and transforming GWAS summary statistics based on the decomposition
}
