% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp.R
\name{prep_RSSp_evd}
\alias{prep_RSSp_evd}
\title{Prepare data for RSSp by transforming GWAS summary statistics based on the precomputed eigenvalue decomposition of the LD matrix}
\usage{
prep_RSSp_evd(Ql, Dl, U)
}
\arguments{
\item{Ql}{A list of square matrices where each column of each matrix is an eigenvector of the LD matrix}

\item{Dl}{a list of vectors where each element of each vector is an eigenvalue}

\item{U}{A vector or matrix of GWAS standardized effect sizes. If `U` is a matrix, there should be one column for each trait.
It is generally a good idea to keep track of the traits by naming the columns}
}
\value{
Returns a dataframe with columns `quh`, which is the transformed effect size,`D` which are the eigenvalues and `fgeneid`
which is the trait name (taken from the column names of `U`)
}
\description{
Prepare data for RSSp by transforming GWAS summary statistics based on the precomputed eigenvalue decomposition of the LD matrix
}
