% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssp-rpackage.R
\docType{package}
\name{RSSp-package}
\alias{RSSp}
\alias{RSSp-package}
\title{RSSp: Polygenic RSS}
\description{
Implementation of the Regression with Summary Statistics (RSS) method using a polygenic (normal) prior.  RSSp uses GWAS summary statistics and a (external) reference LD panel to estiamte heritability for quantitative traits
}
