% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp-package.r, R/RSSp.R
\docType{package}
\name{RSSp}
\alias{RSSp}
\alias{RSSp-package}
\alias{RSSp}
\title{RSSp.
Polygenic RSS}
\usage{
RSSp(fgeneid = NULL, D, quh, sigu_bounds = c(1e-07, 2.6), a_bounds = c(0,
  1), p_n, doConfound = T, log_params = F, useGradient = T)
}
\arguments{
\item{fgeneid}{a character vector specifying the name of the traits being analyzed}

\item{D}{A vector of length `p` indicating the eigenvalues obtained from an EVD of the LD matrix.}

\item{quh}{a `p` by `ntrait` matrix with transformed, standardized effect sizes obtained from running `prep_RSSp`
(If only one trait is being analyzed, `quh` can also be a length `p` vector)}

\item{sigu_bounds}{a length two vector indicating bounds in the parameter `sigu` to use when optimizing `sigu` is directly related to PVE}

\item{a_bounds}{a length two vector indicating the bounds of the parameter `a` to use when optimizing, `a` is a measure of confounding}

\item{doConfound}{a logical indicating whether to use the two parameter model (`doConfound=T`) or the one parameter model without confounding}

\item{log_params}{a logical indicating whether to optimize parameters in log space or not (not recommended)}
}
\description{
RSSp.
Polygenic RSS

Run RSSp on one or several traits
}
