% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp.R
\name{RSSp_cov}
\alias{RSSp_cov}
\title{Estimate the quantile of the maximum likelihood estimate}
\usage{
RSSp_cov(par0, dvec, quh, perc = 0.95)
}
\description{
Estimate the quantile of the maximum likelihood estimate
}
