% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp.R
\name{RSSp_estimate}
\alias{RSSp_estimate}
\title{Estimate heritability using RSSp}
\usage{
RSSp_estimate(quh, D, sample_size = NULL, trait_id = gen_trait_uuid(),
  pve_bounds = c(.Machine$double.eps, 1 - .Machine$double.eps),
  nterms = 1, eigenvalue_cutoff = 0.001, calc_H = F, alt_pve = F)
}
\description{
Estimate heritability using RSSp
}
