% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp.R
\name{RSSp_information_nc}
\alias{RSSp_information_nc}
\title{Estimate the quantile of the maximum likelihood estimate
Title}
\usage{
RSSp_information_nc(par, dvec)
}
\arguments{
\item{quh}{}
}
\description{
Estimate the quantile of the maximum likelihood estimate
Title
}
