% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_beta_gds}
\alias{sim_beta_gds}
\title{Generate simulated values of beta, given PVE etc.}
\usage{
sim_beta_gds(gds, tparam_df, seed = NULL, chunksize = 10000)
}
\arguments{
\item{gds}{an open SeqArray gds file handle}

\item{tparam_df}{A dataframe with the simulation parameter values (see `gen_tparamdf_norm`)}

\item{seed}{a random seed}

\item{chunksize}{number of values to create at a time}
}
\description{
Generate simulated values of beta, given PVE etc.
}
