% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_quh_dir}
\alias{sim_quh_dir}
\title{"Directly" simulate data from the RSSp likelihood for a single value of `sigu` and `bias`.}
\usage{
sim_quh_dir(tsigu, tbias, fgeneid = NULL, Q, D, seed = NULL, snp_id)
}
\arguments{
\item{tsigu}{desired sd of the true effect}

\item{tbias}{true value for bias/confounding}

\item{fgeneid}{name for the trait (character)}

\item{Q}{matrix of eigenvectors of the LD matrix (must be square, and must have rank equal to nrow(usim))}

\item{D}{vector of eigenvalues corresponding to Q ( must have length equal to rank of `Q`)}

\item{snp_id}{vector of IDs for the SNPs}
}
\description{
"Directly" simulate data from the RSSp likelihood for a single value of `sigu` and `bias`.
}
