% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_quh_dir_df}
\alias{sim_quh_dir_df}
\title{"Directly" simulate data from the RSSp likelihood for a range of values of sigu and bias}
\usage{
sim_quh_dir_df(tparam_df, Q, D, seed = NULL, snp_id)
}
\arguments{
\item{Q}{matrix of eigenvectors of the LD matrix (must be square, and must have rank equal to nrow(usim))}

\item{D}{vector of eigenvalues corresponding to Q ( must have length equal to rank of `Q`)}

\item{snp_id}{vector of IDs for the SNPs}

\item{tsigu}{desired sd of the true effect (vector of length `g``, where `g` is the number of traits to be simulated)}

\item{tbias}{true value for bias/confounding}

\item{fgeneid}{name for the trait (character)}
}
\description{
"Directly" simulate data from the RSSp likelihood for a range of values of sigu and bias
}
