% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{sim_uh_A}
\alias{sim_uh_A}
\title{Transform simulated standard univariate normal data to multivariate normal, following the RSSp likelihood}
\usage{
sim_uh_A(sigu, bias, Q, D, fgeneid, usim)
}
\arguments{
\item{sigu}{desired sd of the true effect}

\item{bias}{true value for bias/confounding}

\item{Q}{matrix of eigenvectors of the LD matrix (must be square, and must have rank equal to nrow(usim))}

\item{D}{vector of eigenvalues corresponding to Q ( must have length equal to rank of `Q`)}

\item{fgeneid}{name for the trait (character)}

\item{usim}{matrix of standard univariate normal data to be transformed, one column for each replicate.}
}
\description{
Transform simulated standard univariate normal data to multivariate normal, following the RSSp likelihood
}
