% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp_prediction.R
\name{simulate_v}
\alias{simulate_v}
\title{Function for simulating "true" orthogonalized effects given eigenvalues and PVE parameters}
\usage{
simulate_v(cvec, D)
}
\description{
Function for simulating "true" orthogonalized effects given eigenvalues and PVE parameters
}
