% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSp_utils.R
\name{transform_D}
\alias{transform_D}
\title{Internal helper function for components of the confounding model}
\usage{
transform_D(cvec, D, i = 1)
}
\arguments{
\item{cvec}{inferred parameters for the model}

\item{D}{vector of eigenvalues of the LD matrix}

\item{i}{integer index for which term in the model to return.  Must be between 1 and length(cvec).}
}
\value{
a vector of length equal to `length(D)` where each element represents the estimated variance contributed by the indicated (i-th) component
}
\description{
Internal helper function for components of the confounding model
}
