% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evd_dnorm}
\alias{evd_dnorm}
\title{evd_dnorm}
\usage{
evd_dnorm(par, dvec, quh)
}
\arguments{
\item{par}{a length 2 numeric vector.  par[1] corresponds to the sigma_u parameter, and par[2] corresponds to the confounding parameter}

\item{quh}{The precomputed matrix vector product Q%*%u_hat (passed as a vector)}

\item{dvec.}{The eigenvalues of the LD matrix}
}
\description{
This function computes the RSSp negative log likelihood
}
